// src/shared/enums/alert-type.enum.ts
export enum AlertType {
  // Críticos
  HIGH_INTERNAL_TEMP = 'HIGH_INTERNAL_TEMP',
  LOW_INTERNAL_TEMP = 'LOW_INTERNAL_TEMP',
  CRITICAL_WEIGHT_LOSS = 'CRITICAL_WEIGHT_LOSS',
  HIVE_OFFLINE = 'HIVE_OFFLINE',

  // Avisos
  SENSOR_FAILURE = 'SENSOR_FAILURE',
  HIGH_INTERNAL_HUMIDITY = 'HIGH_INTERNAL_HUMIDITY',
  LOW_INTERNAL_HUMIDITY = 'LOW_INTERNAL_HUMIDITY',

  // Informações
  HARVEST_READY = 'HARVEST_READY',
}